rag_bot/requirements.txt
sentence-transformers
faiss-cpu
pymupdf
numpy
accelerate
transformers>=4.30.0
torch>=2.0.0
safetensors
gradio==3.36.0